backend/requirements.txt
django
djangorestframework
psycopg2-binary
Pillow
django-filter
django-cors-headers

requests
django-extensions